[OPIK-5333] [BE] fix: redact span scorer log and unblock CI
Move summarizeEvaluatorInput to OnlineScoringEngine and use it from the
span scorer too — the trace scorer was already redacted, the span side
was still dumping the full rendered map. Widen the integration test
matcher in AutomationRuleEvaluatorsResourceTest to accept the new
'Sending traceId ... to Python evaluator: arguments=[...]' summary
format alongside the LLM scorer's summary.
